requirements.txt
pandas
Faker
kafka-python